requirements.txt
# Core dependencies
openai>=1.0.0
anthropic>=0.5.0
python-dotenv>=1.0.0

# Data processing
pandas>=2.2.0
numpy>=2.0.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0

# Development
pytest>=7.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0

# Added from the code block
matplotlib>=3.10.0
google-generativeai>=0.3.0 